pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "billboardsingles"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "beautifulsoup4==4.12.2",
    "openai==0.27.8",
    "pandas==2.0.2",
    "python-dotenv==1.0.0",
    "prefect==2.10.12",
    "requests==2.31.0",
    "tqdm==4.65",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
[project.optional-dependencies]
tests = [
  "pytest",
  "black",
]

[project.urls]
"Homepage" = "https://github.com/jandolezal/billboard"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
